Move background_2d interpolators to new module
